Bullet collison detection

+++ OLC game jam 2019/Player.h
@@ -23,7 +23,7 @@ public:
 	virtual void Update(float delta_time);
 	virtual float GetX() { return pos_x;  }
 	virtual float GetY() { return pos_y; }
-	virtual float GetWidth() { return width;  }
-	virtual float GetHeight() { return height; }
+	virtual float GetWidth() { return static_cast<float>(width);  }
+	virtual float GetHeight() { return static_cast<float>(height); }
 };
